Fix shouldComponentUpdate in OverflowList
Original expression: `!(this.state !== nextState && shallowCompareKeys(this.state, nextState))`
With DeMorgans laws can be written as `(this.state === nextState || !shallowCompareKeys(this.state, nextState))`

This is saying that the component SHOULD update iff: the states are equal by reference OR the keys are NOT shallow equal. 

I don't believe this was the intended logic, based on the above comments -- I imagine we don't actually want to re-render if they states have referential equality.